Changes to HandsTesting, MenuToggle, MenuFunctons, ButtonScript, and Ring Toss.
Ring Toss is just minor positioning changes, but a lot of them.
HandsTesting added a section for punching.
Hands testing is working on a HMDRotatesY player.
Hands testing has a hand menu that toggles HMD rotatesY.
Hands testing has a working Key and door. Get the key fast. It tends to roll off.